WIP: First attempt at splitting JNM
